Authentication/requirements.txt
bcrypt==3.2.0
cffi==1.14.3
click==7.1.2
cryptography==3.2.1
databases==0.4.0
dnspython==2.0.0
ecdsa==0.14.1
fastapi==0.61.1
h11==0.11.0
loguru==0.5.3
marshmallow==2.21.0
motor==2.3.0
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.7.1
pymongo==3.11.0
python-dateutil==2.8.1
python-jose==3.2.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.20
starlette==0.13.6
typing-extensions==3.7.4.3
umongo==2.3.0
uvicorn==0.12.2
